+++ src/lib.rs
@@ -23,18 +23,18 @@
 //! | Bzip2 | ✅ | ✅ |
 //! | ZStandard | ✅ | ✅ |
 //! | LZMA | ✅ | |
-//! | XZ | ✅ | |
+//! | XZ | ✅ | ✅ |
 //! | PPMd | ✅ | ✅ |
 //! | AES encryption | ✅ | ✅ |
 //! | ZipCrypto deprecated encryption | ✅ | ✅ |
 //!
 //!
-#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![warn(missing_docs)]
 #![allow(unexpected_cfgs)] // Needed for cfg(fuzzing) on nightly as of 2024-05-06
 pub use crate::compression::{CompressionMethod, SUPPORTED_COMPRESSION_METHODS};
 pub use crate::read::HasZipMetadata;
-pub use crate::read::ZipArchive;
+pub use crate::read::{ZipArchive, ZipReadOptions};
 pub use crate::spec::{ZIP64_BYTES_THR, ZIP64_ENTRY_THR};
 pub use crate::types::{AesMode, DateTime};
 pub use crate::write::ZipWriter;
@@ -55,6 +55,8 @@ mod types;
 pub mod write;
 mod zipcrypto;
 pub use extra_fields::ExtraField;
+#[cfg(feature = "legacy-zip")]
+mod legacy;
 
 #[doc = "Unstable APIs\n\
 \